extract contacts styles
