+++ index.html
@@ -61,9 +61,10 @@
         </div>
 
         <div id="lobby" class="lobby">
-          <button class="lobby-btn" id="lobby-call-btn">
+          <button title="Start Call" class="lobby-btn" id="lobby-call-btn">
             <i class="fa fa-phone"></i>
           </button>
+
           <!-- <p id="lobby-text" class="p">
             Or create a new link and share it with a friend:
           </p>
@@ -75,7 +76,7 @@
             </button>
             <button id="copy-link-btn" title="Copy chat link" disabled>
               <i class="fa fa-copy"></i>
-            </button> 
+            </button>
           </div> -->
         </div>
       </div>
