+++ src/firebase/watch-sync.js
@@ -18,7 +18,7 @@ import {
 
 import { sharedVideoEl, sharedBoxEl } from '../elements.js'; // TODO: refactor:
 import { hideElement, showElement } from '../utils/ui/ui-utils.js';
-import { enterWatchMode } from '../main.js';
+import { enterWatchMode } from '../components/ui/watch-mode.js';
 
 // ============================================================================
 // WATCH-TOGETHER SYNC (Firebase-based)
